refactor(npm-scripts): extract createFederationConfig to a module
This is to separate tweakWebpackConfig from the creation of the federation
configuration, which are two separate logical processes. The same reasons
as in previous commit apply.
